requirement.txt
fastapi==0.109.1
uvicorn==0.27.0
sqlalchemy==2.0.25
python-dotenv==1.0.0
pydantic==2.5.3
email-validator==2.0.0.post2
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.6
pytest==7.4.0
httpx==0.25.0